Consolidate the iOptron mount functions into a base class. Other mount cleanup.
